feat(middleware): add separate access log middleware with config option
