+++ src/middleware/whitelist.js
@@ -10,9 +10,6 @@ import { color, getConfigValue, safeReadFileSync } from '../util.js';
 const whitelistPath = path.join(process.cwd(), './whitelist.txt');
 const enableForwardedWhitelist = getConfigValue('enableForwardedWhitelist', false);
 let whitelist = getConfigValue('whitelist', []);
-let knownIPs = new Set();
-
-export const getAccessLogPath = () => path.join(globalThis.DATA_ROOT, 'access.log');
 
 if (fs.existsSync(whitelistPath)) {
     try {
@@ -48,30 +45,12 @@ function getForwardedIp(req) {
     return undefined;
 }
 
-export function migrateAccessLog() {
-    try {
-        if (!fs.existsSync('access.log')) {
-            return;
-        }
-        const logPath = getAccessLogPath();
-        if (fs.existsSync(logPath)) {
-            return;
-        }
-        fs.renameSync('access.log', logPath);
-        console.log(color.yellow('Migrated access.log to new location:'), logPath);
-    } catch (e) {
-        console.error('Failed to migrate access log:', e);
-        console.info('Please move access.log to the data directory manually.');
-    }
-}
-
 /**
  * Returns a middleware function that checks if the client IP is in the whitelist.
  * @param {boolean} whitelistMode If whitelist mode is enabled via config or command line
- * @param {boolean} listen If listen mode is enabled via config or command line
  * @returns {import('express').RequestHandler} The middleware function
  */
-export default function whitelistMiddleware(whitelistMode, listen) {
+export default function whitelistMiddleware(whitelistMode) {
     const forbiddenWebpage = Handlebars.compile(
         safeReadFileSync('./public/error/forbidden-by-whitelist.html') ?? '',
     );
@@ -80,21 +59,6 @@ export default function whitelistMiddleware(whitelistMode, listen) {
         const clientIp = getIpFromRequest(req);
         const forwardedIp = getForwardedIp(req);
         const userAgent = req.headers['user-agent'];
-
-        if (listen && !knownIPs.has(clientIp)) {
-            console.info(color.yellow(`New connection from ${clientIp}; User Agent: ${userAgent}\n`));
-            knownIPs.add(clientIp);
-
-            // Write access log
-            const logPath = getAccessLogPath();
-            const timestamp = new Date().toISOString();
-            const log = `${timestamp} ${clientIp} ${userAgent}\n`;
-            fs.appendFile(logPath, log, (err) => {
-                if (err) {
-                    console.error('Failed to write access log:', err);
-                }
-            });
-        }
 
         //clientIp = req.connection.remoteAddress.split(':').pop();
         if (whitelistMode === true && !whitelist.some(x => ipMatching.matches(clientIp, ipMatching.getMatch(x)))
